+++ core-ng/src/main/java/core/framework/impl/module/Config.java
@@ -3,8 +3,9 @@ package core.framework.impl.module;
 /**
  * @author neo
  */
-public abstract class Config {  // use abstract class not interface is for hiding those method from app module
+public abstract class Config {  // use abstract class not interface is for hiding those method from app module, interface methods must be public
     protected abstract void initialize(ModuleContext context, String name);
 
-    protected abstract void validate();
+    protected void validate() {
+    }
 }