+++ core-ng/src/main/java/core/framework/module/SchedulerConfig.java
@@ -37,10 +37,6 @@ public final class SchedulerConfig extends Config {
         context.route(HTTPMethod.POST, "/_sys/job/:job", schedulerController::triggerJob, true);
     }
 
-    @Override
-    protected void validate() {
-    }
-
     public void timeZone(ZoneId zoneId) {
         if (triggerAdded) throw new Error("schedule timeZone must be configured before adding trigger");
         if (zoneId == null) throw new Error("zoneId must not be null");