+++ core-ng/src/test/java/core/framework/module/HTTPConfigTest.java
@@ -1,8 +1,7 @@
 package core.framework.module;
 
-import core.framework.http.HTTPMethod;
 import core.framework.impl.module.ModuleContext;
-import core.framework.impl.web.HTTPServerIOHandler;
+import core.framework.impl.web.HTTPIOHandler;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
@@ -10,34 +9,41 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
+import static core.framework.http.HTTPMethod.DELETE;
+import static core.framework.http.HTTPMethod.GET;
+import static core.framework.http.HTTPMethod.PATCH;
+import static core.framework.http.HTTPMethod.POST;
+import static core.framework.http.HTTPMethod.PUT;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * @author neo
  */
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class RouteConfigTest {
-    private RouteConfig config;
+class HTTPConfigTest {
+    private HTTPConfig config;
 
     @BeforeAll
     void createRouteConfig() {
-        config = new RouteConfig(new ModuleContext());
+        config = new HTTPConfig();
+        config.initialize(new ModuleContext(), null);
     }
 
     @Test
     void route() {
-        TestControllers controllers = new TestControllers();
-        config.get("/route-test", controllers::get);
-        config.post("/route-test", controllers::post);
-        TestController controller = new TestController();
-        config.put("/route-test", controller);
-        config.delete("/route-test", controller);
-        config.patch("/route-test", controller);
+        var controllers = new TestControllers();
+        config.route(GET, "/route-test", controllers::get);
+        config.route(POST, "/route-test", controllers::post);
+
+        var controller = new TestController();
+        config.route(PUT, "/route-test", controller);
+        config.route(DELETE, "/route-test", controller);
+        config.route(PATCH, "/route-test", controller);
     }
 
     @Test
     void routeWithReservedPath() {
-        assertThatThrownBy(() -> config.add(HTTPMethod.GET, HTTPServerIOHandler.HEALTH_CHECK_PATH, new TestController()))
+        assertThatThrownBy(() -> config.route(GET, HTTPIOHandler.HEALTH_CHECK_PATH, new TestController()))
                 .isInstanceOf(Error.class)
                 .hasMessageContaining("/health-check is reserved path");
     }